+++ source/router/search/src/vscode.rs
@@ -1,3 +1,5 @@
+use std::fmt::Write;
+
 use crate::components::SearchState;
 use flams_ontology::{
     search::{QueryFilter, SearchResult, SearchResultKind},
@@ -168,7 +170,11 @@ fn do_results(
     remote: Option<fn() -> Option<String>>,
     results: RwSignal<SearchState>,
 ) -> impl IntoView {
-    use leptos::either::EitherOf5::*;
+    use leptos::either::EitherOf6::*;
+    inject_css(
+        "ftml-comp",
+        include_str!("../../../ftml/viewer-components/src/components/comp.css"),
+    );
     let pre_view =
         move || view! {<div style="width:100%;font-weight:bold;text-align:center;">{pre}</div>};
     move || {
@@ -176,11 +182,15 @@ fn do_results(
             SearchState::None => A(()),
             SearchState::Results(v) if v.is_empty() => B(view!({pre_view}"(No results)")),
             SearchState::Loading => C(view!({pre_view}<flams_web_utils::components::Spinner/>)),
-            SearchState::SymResults(v) => D(view!({pre_view}{v
+            SearchState::SymResults(v) if remote.is_none() => D(view!({pre_view}{v
             .iter()
-            .map(|(sym, res)| do_sym_result(sym, res.clone(),remote))
+            .map(|(sym, _)| do_sym_result_local(sym))
             .collect_view()})),
-            SearchState::Results(v) => E(view!({pre_view}{v
+            SearchState::SymResults(v) => E(view!({pre_view}{v
+            .iter()
+            .map(|(sym, res)| do_sym_result_remote(sym, res.clone(),unwrap!(remote)))
+            .collect_view()})),
+            SearchState::Results(v) => F(view!({pre_view}{v
             .iter()
             .map(|(score, res)| do_result(*score, res,remote))
             .collect_view()})),
@@ -204,6 +214,7 @@ fn do_result(
 
 #[derive(leptos::server_fn::serde::Serialize, Debug, Clone)]
 struct Usemodule {
+    kind: &'static str,
     archive: ArchiveId,
     path: String,
 }
@@ -216,38 +227,85 @@ impl Usemodule {
         } else {
             module.name().first_name().to_string()
         };
-        Self { archive, path }
+        Self {
+            kind: "usemodule",
+            archive,
+            path,
+        }
     }
 }
 
-fn do_sym_result(
-    sym: &SymbolURI,
-    res: Vec<(f32, SearchResult)>,
-    remote: Option<fn() -> Option<String>>,
-) -> impl IntoView + use<> {
-    use thaw::Scrollbar;
+#[derive(leptos::server_fn::serde::Serialize, Debug, Clone)]
+struct Preview<'u> {
+    kind: &'static str,
+    uri: &'u SymbolURI,
+}
+impl Preview<'_> {
+    fn make(uri: &SymbolURI) -> Preview<'_> {
+        Preview {
+            kind: "preview",
+            uri,
+        }
+    }
+}
+
+#[derive(Copy, Clone)]
+struct Short<'u>(&'u SymbolURI);
+impl std::fmt::Display for Short<'_> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "[{}]{{", self.0.archive_id())?;
+        if let Some(p) = self.0.path() {
+            p.fmt(f)?;
+            f.write_char('?')?;
+        }
+        write!(f, "{}}} {}", self.0.module().name(), self.0.name())
+    }
+}
+
+fn do_sym_result_local(sym: &SymbolURI) -> impl IntoView + use<> {
     let vs = unwrap!(VSCode::get());
-    let usemodule = if remote.is_none() {
-        Some(Usemodule::make(sym))
-    } else {
-        None
-    };
-    let name = ftml_viewer_components::components::omdoc::symbol_name(sym, &sym.to_string());
+    let name = ftml_viewer_components::components::omdoc::symbol_name(sym, &Short(sym).to_string());
     view! {
         <div class="flams-search-block">
             <div><b>{name}</b>
                 {
-                    usemodule.map(|u| {
-                        let on_click = move |_| {
-                            let _ = vs.post_message(u.clone());
-                        };
-                        view!{
-                            <div style="width:100%"><div style="margin-left:auto;width:fit-content;" on:click=on_click>
-                                <VSCodeButton>"\\usemodule"</VSCodeButton>
-                            </div></div>
-                        }
-                    })
+                    let sym_a = sym.clone();
+                    let vs_a = vs.clone();
+                    let on_use = move |_| {
+                        let _ = vs_a.post_message(Usemodule::make(&sym_a));
+                    };
+                    let sym = sym.clone();
+                    let on_preview = move |_| {
+                        let _ = vs.post_message(Preview::make(&sym));
+                    };
+                    view!{
+                        <div style="width:100%">
+                            <div style="margin-left:auto;width:fit-content;display:flex;flex-direction:row;">
+                                <div style="width:fit-content;margin-right:5px;" on:click=on_preview>
+                                    <VSCodeButton>"preview"</VSCodeButton>
+                                </div>
+                                <div style="width:fit-content;" on:click=on_use>
+                                    <VSCodeButton>"\\usemodule"</VSCodeButton>
+                                </div>
+                            </div>
+                        </div>
+                    }
                 }
+            </div>
+        </div>
+    }
+}
+
+fn do_sym_result_remote(
+    sym: &SymbolURI,
+    res: Vec<(f32, SearchResult)>,
+    remote: fn() -> Option<String>,
+) -> impl IntoView + use<> {
+    use thaw::Scrollbar;
+    let name = ftml_viewer_components::components::omdoc::symbol_name(sym, &sym.to_string());
+    view! {
+        <div class="flams-search-block">
+            <div><b>{name}</b>
             </div>
             <div style="display:block">
             <div style="padding:0 5px;max-width:100%">
@@ -255,7 +313,7 @@ fn do_sym_result(
                   <Scrollbar style="max-height: 100px;width:100%;max-width:100%;">{
                     res.into_iter().map(|(_,r)| {
                       let SearchResult::Paragraph { uri, .. } = r else { impossible!()};
-                      fragment(uri.into(),remote)
+                      fragment(uri.into(),Some(remote))
                     }).collect_view()
                   }
                   </Scrollbar>